Clarify sandbox hook logic

+++ src/sandbox.sh
@@ -384,22 +384,22 @@ function build_sandbox () {
 		log 'Starting to build sandbox layer'
 	fi
 
-	if ! [ -d "${HALCYON_DIR}/sandbox" ]; then
-		if [ -f "${app_dir}/.halcyon-magic/sandbox-precreate-hook" ]; then
-			log 'Running sandbox pre-create hook'
-			( "${app_dir}/.halcyon-magic/sandbox-precreate-hook" "${ghc_tag}" "${sandbox_tag}" "${app_dir}" ) || die
-			mkdir -p "${HALCYON_DIR}/sandbox/.halcyon-magic" || die
-			cp "${app_dir}/.halcyon-magic/sandbox-precreate-hook" "${HALCYON_DIR}/sandbox/.halcyon-magic" || die
-		fi
+	if ! (( ${extending_sandbox} )) && [ -f "${app_dir}/.halcyon-magic/sandbox-precreate-hook" ]; then
+		log 'Running sandbox pre-create hook'
+		( "${app_dir}/.halcyon-magic/sandbox-precreate-hook" "${ghc_tag}" "${sandbox_tag}" "${app_dir}" ) || die
+		mkdir -p "${HALCYON_DIR}/sandbox/.halcyon-magic" || die
+		cp "${app_dir}/.halcyon-magic/sandbox-precreate-hook" "${HALCYON_DIR}/sandbox/.halcyon-magic" || die
+	fi
 
+	if ! (( ${extending_sandbox} )) && ! [ -d "${HALCYON_DIR}/sandbox" ]; then
 		cabal_create_sandbox "${HALCYON_DIR}/sandbox" || die
+	fi
 
-		if [ -f "${app_dir}/.halcyon-magic/sandbox-postcreate-hook" ]; then
-			log 'Running sandbox post-create hook'
-			( "${app_dir}/.halcyon-magic/sandbox-postcreate-hook" "${ghc_tag}" "${sandbox_tag}" "${app_dir}" ) || die
-			mkdir -p "${HALCYON_DIR}/sandbox/.halcyon-magic" || die
-			cp "${app_dir}/.halcyon-magic/sandbox-postcreate-hook" "${HALCYON_DIR}/sandbox/.halcyon-magic" || die
-		fi
+	if ! (( ${extending_sandbox} )) && [ -f "${app_dir}/.halcyon-magic/sandbox-postcreate-hook" ]; then
+		log 'Running sandbox post-create hook'
+		( "${app_dir}/.halcyon-magic/sandbox-postcreate-hook" "${ghc_tag}" "${sandbox_tag}" "${app_dir}" ) || die
+		mkdir -p "${HALCYON_DIR}/sandbox/.halcyon-magic" || die
+		cp "${app_dir}/.halcyon-magic/sandbox-postcreate-hook" "${HALCYON_DIR}/sandbox/.halcyon-magic" || die
 	fi
 
 	if [ -f "${app_dir}/.halcyon-magic/sandbox-prebuild-hook" ]; then
